+++ flake.nix
@@ -1,0 +1,59 @@
+{
+  description = "servant-event-stream";
+
+  inputs = {
+    nixpkgs.url = "github:nixos/nixpkgs/nixos-24.05";
+  };
+
+  outputs =
+    { self, nixpkgs, ... }:
+    let
+      supportedSystems = [ "x86_64-linux" ];
+      forallSystems =
+        f:
+        nixpkgs.lib.genAttrs supportedSystems (
+          system:
+          f (rec {
+            inherit system;
+            pkgs = nixpkgsFor system;
+            haskellPackages = hpkgsFor system pkgs;
+          })
+        );
+      nixpkgsFor = system: import nixpkgs { inherit system; };
+      hpkgsFor =
+        system: pkgs:
+        with pkgs.haskell.lib;
+        pkgs.haskell.packages.ghc98.override { overrides = self: super: { }; };
+    in
+    {
+      packages = forallSystems (
+        {
+          system,
+          pkgs,
+          haskellPackages,
+        }:
+        {
+          servant-event-stream = haskellPackages.callCabal2nix "servant-event-stream" ./. { };
+          default = self.packages.${system}.servant-event-stream;
+        }
+      );
+      devShells = forallSystems (
+        {
+          system,
+          pkgs,
+          haskellPackages,
+        }:
+        {
+          servant-event-stream = haskellPackages.shellFor {
+            packages = p: [ self.packages.${system}.servant-event-stream ];
+            buildInputs = with haskellPackages; [
+              cabal-install
+              haskell-language-server
+            ];
+            withHoogle = true;
+          };
+          default = self.devShells.${system}.servant-event-stream;
+        }
+      );
+    };
+}