+++ shell.nix
@@ -1,1 +1,0 @@
-(import ./servant-event-stream.nix).shell