+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Revision history for servant-event-stream
 
+## 0.3.0.0 -- 2024-09-05
+
+* Breaking changes to the API.
+
+    Event streams are implemented using servant's 'Stream' endpoint. You should
+    provide a handler that returns a stream of events that implements 'ToSourceIO'
+    where events have a 'ToServerEvent' instance.
+
+    Example:
+
+    > type MyApi = "books" :> ServerSentEvents (SourceIO Book)
+    >
+    > instance ToServerEvent Book where
+    >   toServerEvent book = ...
+    >
+    > server :: Server MyApi
+    > server = streamBooks
+    >   where streamBooks :: Handler (SourceIO Book)
+    >         streamBooks = pure $ source [book1, ...]
+
 ## 0.2.1.0 -- 2021-04-21
 
 * Import `Data.Semigroup` for base < 4.11.0